% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountsEntropyPrefunctions.R
\name{annuli_counts}
\alias{annuli_counts}
\title{Convert Cumulative Counts in Concentric Polygons to Counts in Annuli Polygons}
\usage{
annuli_counts(mat)
}
\arguments{
\item{mat}{Matrix containing cumulative counts.}
}
\value{
A matrix with updated counts.
}
\description{
Converts cumulative counts of concentric polygons to actual counts in annuli polygons.
}
