% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_MCC.R
\name{are_neighbors}
\alias{are_neighbors}
\title{Determine if Cells are Neighbors}
\usage{
are_neighbors(df1, df2, threshold)
}
\arguments{
\item{df1}{A data frame with x and y coordinates of the first set of cells.}

\item{df2}{A data frame with x and y coordinates of the second set of cells.}

\item{threshold}{A numeric value specifying the distance threshold to consider cells as neighbors.}
}
\value{
TRUE if any cell in df1 is a neighbor of any cell in df2 within the specified threshold, otherwise FALSE.
}
\description{
Determine if Cells are Neighbors
}
