% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreMethodFunctions.R
\name{calculate_centroids}
\alias{calculate_centroids}
\title{Calculate Centroids}
\usage{
calculate_centroids(df_boundary)
}
\arguments{
\item{df_boundary}{A data frame containing boundary data.}
}
\value{
A data frame containing centroids for each segment.
}
\description{
Calculates centroids for given boundaries.
}
