% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_Entropy.R
\name{calculate_entropy}
\alias{calculate_entropy}
\title{Calculate entropy from a probability vector}
\usage{
calculate_entropy(p)
}
\arguments{
\item{p}{A numeric vector of probabilities.}
}
\value{
The calculated entropy value.
}
\description{
Calculate entropy from a probability vector
}
