% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_MCC.R
\name{calculate_mcc_for_cell}
\alias{calculate_mcc_for_cell}
\title{Calculate MCC for all pairs of genes for a given cell and its neighbors.}
\usage{
calculate_mcc_for_cell(cell_col, neighbor_col)
}
\arguments{
\item{cell_col}{Column of a matrix representing a cell.}

\item{neighbor_col}{Column of a matrix representing a cell's neighbor.}
}
\value{
A vector containing MCC values for gene pairs.
}
\description{
Calculate MCC for all pairs of genes for a given cell and its neighbors.
}
