% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_MCC.R
\name{calculate_neighbour_sums}
\alias{calculate_neighbour_sums}
\title{Calculate Sum of Gene Counts from Neighbors}
\usage{
calculate_neighbour_sums(mat, cellneighbours)
}
\arguments{
\item{mat}{A matrix of gene counts.}

\item{cellneighbours}{A data frame or list with neighbor information.}
}
\value{
A matrix with original gene counts and additional columns for neighbor sums.
}
\description{
Calculate Sum of Gene Counts from Neighbors
}
