% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialFeatures_utils.R
\name{check_if_me_boundaries}
\alias{check_if_me_boundaries}
\title{Check that a MoleculeExperiment object is both valid and contains boundaries}
\usage{
check_if_me_boundaries(me)
}
\arguments{
\item{me}{A MoleculeExperiment object}
}
\value{
Will return an error if me is not valid and containing boundaries
}
\description{
Check that a MoleculeExperiment object is both valid and contains boundaries
}
