% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSE.R
\name{combineSE}
\alias{combineSE}
\title{Combine SummarizedExperiment Objects}
\usage{
combineSE(se_list)
}
\arguments{
\item{se_list}{A list of SummarizedExperiment objects to be combined.
All objects should have an assay named "counts". The function
assumes that all colData DataFrames can be combined with rbind,
and if they have different columns, they need to be harmonized beforehand.}
}
\value{
A SummarizedExperiment object that represents the combination of all
input objects. The returned object contains a unified feature set
from the object with the most features, combined assay data, and
concatenated colData reflecting metadata for each sample across
the combined dataset.
}
\description{
This function combines multiple SummarizedExperiment objects into a single
SummarizedExperiment object. It aligns features (rows) across all objects
based on the object with the most features, fills missing feature data with NA,
and concatenates colData from all objects.
}
\examples{
# Assuming you have a list of SummarizedExperiment objects: se_list
combined_se <- combineSE(se_list)

}
