% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountsEntropyPrefunctions.R
\name{combo_counts}
\alias{combo_counts}
\title{Convert Cumulative Counts in Concentric Polygons to Counts in Combo Polygons}
\usage{
combo_counts(mat)
}
\arguments{
\item{mat}{Matrix containing cumulative counts.}
}
\value{
A matrix with updated counts.
}
\description{
Converts cumulative counts of concentric polygons to actual counts in combo polygons.
}
