% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBoundariesToMEforAll_utils.R
\name{compute_central_coordinates}
\alias{compute_central_coordinates}
\title{Compute Central Coordinates of Boundaries}
\usage{
compute_central_coordinates(me, assay_name = "cell")
}
\arguments{
\item{me}{A Molecule Experiment (ME) object that holds boundary data.}

\item{assay_name}{The name of the assay to extract boundaries from, defaults to "cell".}
}
\value{
A dataframe with segment_id, sample_id, x_central, and y_central columns.
}
\description{
This function computes the central coordinates for each segment in the given
Molecule Experiment (ME) object. The function takes advantage of the \code{boundaries}
function to extract boundary coordinates and then computes the central coordinate
by averaging the x and y locations.
}
\examples{
\dontrun{
# Assuming `me_obj` is your Molecule Experiment object
central_coordinates <- compute_central_coordinates(me_obj, assay_name = "cell")
}
}
