% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountsEntropyPrefunctions.R
\name{compute_entropy}
\alias{compute_entropy}
\title{Compute Entropy for Matrix}
\usage{
compute_entropy(mat)
}
\arguments{
\item{mat}{Matrix for which to compute entropy.}
}
\value{
A data frame of entropy values for each row of the matrix.
}
\description{
Computes entropy for each row of a matrix.
}
