% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_MCC.R
\name{compute_mcc}
\alias{compute_mcc}
\title{Compute Matthews Correlation Coefficient (MCC) for a given 2x2 table.}
\usage{
compute_mcc(table)
}
\arguments{
\item{table}{A 2x2 matrix or table.}
}
\value{
The computed MCC value.
}
\description{
Compute Matthews Correlation Coefficient (MCC) for a given 2x2 table.
}
