% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCCMatrix_utils.R
\name{compute_mcc_matrix}
\alias{compute_mcc_matrix}
\title{Compute MCC Matrix for Spatial Cell Data}
\usage{
compute_mcc_matrix(mat)
}
\arguments{
\item{mat}{A matrix with gene counts for cells and their neighbors.}
}
\value{
A matrix with Matthews correlation coefficients for gene pairs across cells.
}
\description{
Compute MCC Matrix for Spatial Cell Data
}
