% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConversionFunction.R
\name{convert_to_long}
\alias{convert_to_long}
\title{Convert to Long Format}
\usage{
convert_to_long(me, assayName)
}
\arguments{
\item{me}{An object containing information about the experiment.}

\item{assayName}{Character string specifying the assay type.}
}
\value{
A data frame in long format.
}
\description{
Converts a given entropy table to a long format.
}
