% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertoLong.R
\name{ConverttoLong}
\alias{ConverttoLong}
\title{Convert Entropy Matrix to Long Format Dataframe}
\usage{
ConverttoLong(me, assayName, ...)
}
\arguments{
\item{me}{A Molecule Experiment object.}

\item{assayName}{A character string specifying the assay type. Supported values include
"sub-sector", "sub-concentric", "sub-combo", "super_sector", "super-concentric", and "super-combo".}
}
\value{
A dataframe in long format with columns "Value", "Cell", "Gene", and "Approach".
}
\description{
This function converts an entropy matrix from a Molecule Experiment object to a long format dataframe.
}
\examples{
\dontrun{
# Assuming `data_obj` is your Molecule Experiment object
long_df = convert_to_long(data_obj, assayName = "sub_sector")
}
}
