% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_MCC.R
\name{count_molecules_from_SE}
\alias{count_molecules_from_SE}
\title{Count molecules in a SpatialExperiment object.}
\usage{
count_molecules_from_SE(me)
}
\arguments{
\item{me}{A SpatialExperiment object.}
}
\value{
A cell object with counted molecules.
}
\description{
Count molecules in a SpatialExperiment object.
}
