% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntropyMatrix_utils.R
\name{countsprop}
\alias{countsprop}
\title{Calculate prop for entropy}
\usage{
countsprop(counts)
}
\arguments{
\item{counts}{a vector of integers}
}
\value{
the proportion
}
\description{
Calculate prop for entropy
}
