% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_Entropy.R
\name{create_scaled_df}
\alias{create_scaled_df}
\title{Create a scaled data frame}
\usage{
create_scaled_df(s, df_circle, k = 8)
}
\arguments{
\item{s}{A numeric scaling factor.}

\item{df_circle}{A data frame to scale.}

\item{k}{A numeric value, defaulting to 8.}
}
\value{
A scaled data frame.
}
\description{
Create a scaled data frame
}
