% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreMethodFunctions.R
\name{create_sectors}
\alias{create_sectors}
\title{Create Sectors}
\usage{
create_sectors(df)
}
\arguments{
\item{df}{A data frame containing the required fields to create sectors.}
}
\value{
A data frame containing sector data.
}
\description{
Creates sectors based on given data.
}
