% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreMethodFunctions.R
\name{create_sectors_for_scaled}
\alias{create_sectors_for_scaled}
\title{Create Sectors for Scaled Data}
\usage{
create_sectors_for_scaled(df, scale)
}
\arguments{
\item{df}{A data frame containing the required fields.}

\item{scale}{Numeric value for scaling.}
}
\value{
A data frame containing sector data based on scaled inputs.
}
\description{
Creates sectors based on scaled data.
}
