% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreMethodFunctions.R
\name{create_sectors_modified}
\alias{create_sectors_modified}
\title{Create Modified Sectors}
\usage{
create_sectors_modified(df)
}
\arguments{
\item{df}{A data frame containing the required fields to create modified sectors.}
}
\value{
A data frame containing modified sector data.
}
\description{
Creates sectors with modifications based on given data.
}
