% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_Entropy.R
\name{delete_inner}
\alias{delete_inner}
\title{Delete columns ending with '0' from a matrix}
\usage{
delete_inner(m)
}
\arguments{
\item{m}{A matrix to process.}
}
\value{
A modified matrix without columns ending in '0'.
}
\description{
Delete columns ending with '0' from a matrix
}
