% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountsEntropyPrefunctions.R
\name{entropy}
\alias{entropy}
\title{Calculate Entropy}
\usage{
entropy(p)
}
\arguments{
\item{p}{Numeric vector representing a probability distribution.}
}
\value{
The entropy of the distribution.
}
\description{
Calculates the entropy for a given probability distribution.
}
