% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntropyMatrix.R
\name{EntropyMatrix}
\alias{EntropyMatrix}
\title{Compute Entropy Matrix for Specified Assay Name}
\usage{
EntropyMatrix(me, assayName)
}
\arguments{
\item{me}{A data object that contains the input data.}

\item{assayName}{A character string specifying the assay name. It must be one
of "sub-sector", "sub-concentric", "sub-combo", "super-concentric", or
"super-combo".}
}
\value{
A matrix containing the entropy data.
}
\description{
This function calculates the entropy matrix for the specified assay name.
}
