% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreMethodFunctions.R
\name{extract_boundaries}
\alias{extract_boundaries}
\title{Extract Boundaries}
\usage{
extract_boundaries(me)
}
\arguments{
\item{me}{An object containing information about the experiment.}
}
\value{
A data frame containing boundary data.
}
\description{
Extracts boundaries from the ME object.
}
