% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreMethodFunctions.R
\name{extract_boundaries_and_centroids}
\alias{extract_boundaries_and_centroids}
\title{Extract Boundaries and Calculate Centroids}
\usage{
extract_boundaries_and_centroids(me)
}
\arguments{
\item{me}{An object containing information about the experiment.}
}
\value{
A list containing boundary and centroid data frames.
}
\description{
A combined function to extract boundaries and calculate centroids from the ME object.
}
