% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntropyMatrix_utils.R
\name{extract_cell_segments}
\alias{extract_cell_segments}
\title{Extract a submatrix corresponding to a segment from a matrix}
\usage{
extract_cell_segments(mat, cell)
}
\arguments{
\item{mat}{A matrix to process.}

\item{cell}{A character string indicating the segment to extract.}
}
\value{
A submatrix of the provided matrix.
}
\description{
Extract a submatrix corresponding to a segment from a matrix
}
