% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCCMatrix_utils.R
\name{extract_data_and_neighbors}
\alias{extract_data_and_neighbors}
\title{Extract Data and Neighbors from SpatialExperiment}
\usage{
extract_data_and_neighbors(me)
}
\arguments{
\item{me}{A SpatialExperiment object.}
}
\value{
A list containing a data frame of boundaries and a data frame of cell neighbors.
}
\description{
Extract Data and Neighbors from SpatialExperiment
}
