% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_Entropy.R
\name{extract_sub_mat}
\alias{extract_sub_mat}
\title{Extract a submatrix corresponding to a segment from a matrix}
\usage{
extract_sub_mat(mat, segment)
}
\arguments{
\item{mat}{A matrix to process.}

\item{segment}{A character string indicating the segment to extract.}
}
\value{
A submatrix of the provided matrix.
}
\description{
Extract a submatrix corresponding to a segment from a matrix
}
