% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateFeatureData_utils.R
\name{generate_scale_factors_all}
\alias{generate_scale_factors_all}
\title{Generate scale factors for all values of k}
\usage{
generate_scale_factors_all(k)
}
\arguments{
\item{k}{A numeric value.}
}
\value{
A vector of scale factors.
}
\description{
Generate scale factors for all values of k
}
