% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBoundaries_utils.R
\name{GenerateFeatureData}
\alias{GenerateFeatureData}
\title{Generate Feature Data Based on Assay Type Combining Various Polygon Assays}
\usage{
GenerateFeatureData(
  me,
  featureTypes = c("sub_sector", "sub_concentric", "super_sector", "super_concentric"),
  k = 5
)
}
\arguments{
\item{me}{The input data object. This object contains spatial information,
typically regarding segments or regions of interest.}

\item{featureTypes}{Character vector containing the feature types to calculate.}

\item{k}{A numeric value indicating the scaling factor for concentric polygons.
Defaults to 5.}
}
\value{
A list containing dataframes for each assay type:
\itemize{
\item \code{sub_sector}: Feature data for sub-sector polygons.
\item \code{sub_concentric}: Feature data for sub-concentric polygons.
\item \code{sub_combo}: Feature data for sub-combo polygons.
\item \code{super_concentric}: Feature data for super-concentric polygons.
\item \code{super_combo}: Feature data for super-combo polygons.
}
}
\description{
This comprehensive function processes the given data (\code{me}) and returns feature
data based on multiple assay types. The supported assays encapsulated within
this function include:
\itemize{
\item Sub-sector polygons
\item Sub-concentric polygons
\item Sub-combo polygons
\item Super-concentric polygons
\item Super-combo polygons
}

These assays represent various spatial subdivisions or transformations, each
addressing unique feature characteristics. The results from each assay are
combined and returned as a list of dataframes.
}
