% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_MCC.R
\name{get_all_neighbors}
\alias{get_all_neighbors}
\title{Get neighbors for all cells and save to dataframe.}
\usage{
get_all_neighbors(df, threshold = 2)
}
\arguments{
\item{df}{Dataframe containing cell data.}

\item{threshold}{Distance threshold to determine if cells are neighbors.}
}
\value{
A dataframe containing cells and their respective neighbors.
}
\description{
Get neighbors for all cells and save to dataframe.
}
