% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCCMatrix_utils.R
\name{get_cell_neighbours}
\alias{get_cell_neighbours}
\title{Function to get the neighbor IDs for a given cell ID}
\usage{
get_cell_neighbours(cell_id)
}
\arguments{
\item{cell_id}{cell ID}
}
\value{
the neighbor IDs
}
\description{
Function to get the neighbor IDs for a given cell ID
}
