% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_MCC.R
\name{get_neighbors}
\alias{get_neighbors}
\title{Get neighbors for a given cell in a dataframe.}
\usage{
get_neighbors(cell_id, df, threshold = 2)
}
\arguments{
\item{cell_id}{ID of the cell to find neighbors for.}

\item{df}{Dataframe containing cell data.}

\item{threshold}{Distance threshold to determine if cells are neighbors.}
}
\value{
A vector containing IDs of the neighboring cells.
}
\description{
Get neighbors for a given cell in a dataframe.
}
