% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCCMatrix_utils.R
\name{get_neighbour_counts}
\alias{get_neighbour_counts}
\title{Function to sum up gene counts from neighbors}
\usage{
get_neighbour_counts(cell_id)
}
\arguments{
\item{cell_id}{cell IDs}
}
\value{
matrix of gene counts from neighbours
}
\description{
Function to sum up gene counts from neighbors
}
