% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountsEntropyPrefunctions.R
\name{get_segments}
\alias{get_segments}
\title{Get Unique Segments}
\usage{
get_segments(mat)
}
\arguments{
\item{mat}{Matrix from which to extract segment names.}
}
\value{
A character vector of unique segments.
}
\description{
Extracts unique segments from the column names of a matrix.
}
