% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBoundaries.R
\name{loadBoundaries}
\alias{loadBoundaries}
\title{Load Boundaries to Molecule Experiment Object}
\usage{
loadBoundaries(me, ...)
}
\arguments{
\item{me}{A MoleculeExperiment (ME) object}

\item{...}{arguments passed to GenerateFeatureData}
}
\value{
An enriched Molecule Experiment (ME) object that includes boundary
information for the supported assays.
}
\description{
This function takes a Molecule Experiment (ME) object as input and enriches
it with boundary data from a variety of assays. The currently supported
assays are:
\itemize{
\item Sub-sector polygons
\item Sub-concentric polygons
\item Super-sector polygons
\item Super-concentric polygons
}

The function internally processes each assay, extracts respective boundary
information, and updates the original ME object with this new data.
}
\examples{
data(example_me)
me <- loadBoundaries(me)
me
}
