% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBoundariesToME.R
\name{loadBoundariesToME}
\alias{loadBoundariesToME}
\title{Load Boundaries to ME Data Object}
\usage{
loadBoundariesToME(me)
}
\arguments{
\item{me}{A data object to which the boundaries will be added.}
}
\value{
The updated ME data object with added boundaries.
}
\description{
This function loads boundaries for multiple assay names into the specified
ME data object.
}
