% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntropySummarizedExperiment_utils.R
\name{make_assay_data}
\alias{make_assay_data}
\title{make assay data}
\usage{
make_assay_data(df_list)
}
\arguments{
\item{df_list}{A list of data frames}
}
\value{
A single data frame
}
\description{
make assay data
}
