% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntropyMatrix_utils.R
\name{matrix_entropy}
\alias{matrix_entropy}
\title{Compute entropy for the entire matrix}
\usage{
matrix_entropy(mat, nCores = 1)
}
\arguments{
\item{mat}{A matrix to compute entropy for.}

\item{nCores}{Number of cores}
}
\value{
A data frame of entropy values for the matrix.
}
\description{
Compute entropy for the entire matrix
}
