% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCCMatrix.R
\name{MCCMatrix}
\alias{MCCMatrix}
\title{Compute the MCC matrix for spatial cell data.}
\usage{
MCCMatrix(me)
}
\arguments{
\item{me}{A SpatialExperiment object.}
}
\value{
A sparse matrix with MCC calculations.
}
\description{
Compute the MCC matrix for spatial cell data.
}
