% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{me}
\alias{me}
\title{Example MoleculeExperiment data}
\format{
A \linkS4class{MoleculeExperiment} object.
}
\usage{
data(example_me, package = 'SpatialFeatures')
}
\description{
A tiny object containing a few cells and molecules
}
\keyword{datasets}
