% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateFeatureData_utils.R
\name{modify_area_id}
\alias{modify_area_id}
\title{Modify the area_id format}
\usage{
modify_area_id(id)
}
\arguments{
\item{id}{A string representing the area_id.}
}
\value{
A string with the modified area_id format.
}
\description{
This function takes an area_id string and modifies the last segment
if it's a single-digit number between 1 and 9 by adding a leading zero.
}
