% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialFeatures-package.R
\docType{package}
\name{SpatialFeatures-package}
\alias{SpatialFeatures}
\alias{SpatialFeatures-package}
\title{SpatialFeatures: Entropy-based subcellular and supercellular features for molecule-resolved spatial omics datasets}
\description{
This package uses molecule-level information to extract new cell-level features as an alternative to simply calculating gene counts. By using four categories, sub-sector, super-sector, super-sector and super-concentric segmentations of cells, SpatialFeatures then uses entropy as a metric to arrive at a cell-by-gene level feature. Overall, this means that we can extract more nuanced information from molecule-resolved spatial gene expression for further downstream analysis with SingleCellExperiment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sydneybiox.github.io/SpatialFeatures/}
  \item Report bugs at \url{https://github.com/sydneybiox/SpatialFeatures/issues}
}

}
\author{
\strong{Maintainer}: Shila Ghazanfar \email{shazanfar@gmail.com} (\href{https://orcid.org/0000-0001-7861-6997}{ORCID}) [contributor]

Other contributors:
\itemize{
  \item Guan Gui [contributor]
}

}
\keyword{internal}
