% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialFeatures.R
\name{spatialFeatures}
\alias{spatialFeatures}
\title{Calculate SpatialFeatures}
\usage{
spatialFeatures(
  me,
  featureTypes = c("sub_sector", "sub_concentric", "super_sector", "super_concentric"),
  k = 5,
  nCores = 1,
  includeCounts = FALSE,
  ...
)
}
\arguments{
\item{me}{A MoleculeExperiment (ME) object}

\item{featureTypes}{a character vector listing the types of featureTypes to include}

\item{k}{A numeric value indicating the number of polygons to calculate entropy (default 5)}

\item{nCores}{number of cores for parallel processing (default 1)}

\item{includeCounts}{logical (default FALSE) whether to include gene counts as features}

\item{...}{arguments passed to loadBoundaries and EntropyMatrix}
}
\value{
A SummarizedExperiment object containing a spatialFeatures assay and cell-level colData
}
\description{
This function takes a MoleculeExperiment (ME) object as input and calculates
new molecule-based spatial features according to the following segmentations:
\itemize{
\item Sub-sector polygons
\item Sub-concentric polygons
\item Super-sector polygons
\item Super-concentric polygons
}

The function has 3 key steps, 1. adds new boundaries to the ME object
according to the subcellular and supercellular segmentations, 2. calculates
entropy for each of the cell and featureType combinations, and 3. combines
the entropy values into a SummarizedExperiment object.
}
\examples{
data(example_me)
se <- spatialFeatures(me)
se
}
