% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_Entropy.R
\name{split_dataframe_by_region}
\alias{split_dataframe_by_region}
\title{Split Rectangle into Regions Based on Central Coordinates}
\usage{
split_dataframe_by_region(df, regions_segments)
}
\arguments{
\item{me}{A Molecule Experiment (ME) object that holds boundary data.}
}
\value{
A dataframe that associates each segment with a region within the rectangle.
}
\description{
This function determines which region each segment belongs to within a rectangle.
It computes the central coordinates and then based on these central points, segments
the rectangle into 12 regions. The function finally associates each segment to one
of the 12 regions.
}
\examples{
\dontrun{
# Assuming `me_obj` is your Molecule Experiment object
segmented_regions <- split_rectangle(me_obj)
}
}
