% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntropyMatrix_utils.R
\name{annuli_counts}
\alias{annuli_counts}
\title{Convert cumulative counts in concentric polygons to annuli polygon counts}
\usage{
annuli_counts(mat)
}
\arguments{
\item{mat}{A matrix of cumulative counts.}
}
\value{
A matrix of annuli polygon counts.
}
\description{
Convert cumulative counts in concentric polygons to annuli polygon counts
}
