% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_MCC.R
\name{are_neighbors}
\alias{are_neighbors}
\title{Check if two cells are neighbors based on their boundary points.}
\usage{
are_neighbors(df1, df2, threshold)
}
\arguments{
\item{df1}{Dataframe of the first cell's boundaries.}

\item{df2}{Dataframe of the second cell's boundaries.}

\item{threshold}{Distance threshold to determine if cells are neighbors.}
}
\value{
Logical indicating if the cells are neighbors.
}
\description{
Check if two cells are neighbors based on their boundary points.
}
