% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindRows.R
\name{BindRows}
\alias{BindRows}
\title{Bind Rows of Multiple Methods into a Single Dataframe}
\usage{
BindRows(me, methods)
}
\arguments{
\item{me}{A data object that contains the input data.}

\item{methods}{A character vector containing the names of the methods to be
processed. They must be among "sub-sector", "sub-concentric", "sub-combo",
"super-concentric", and "super-combo".}
}
\value{
A dataframe in long format.
}
\description{
This function binds rows of multiple processing methods into a single
dataframe in long format.
}
