% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateFeatureData_utils.R
\name{calculate_centroids}
\alias{calculate_centroids}
\title{Calculate centroids from boundaries}
\usage{
calculate_centroids(df_boundary)
}
\arguments{
\item{df_boundary}{A data frame of boundaries.}
}
\value{
A data frame of centroids.
}
\description{
Calculate centroids from boundaries
}
