% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_MCC.R
\name{calculate_neighbour_sums}
\alias{calculate_neighbour_sums}
\title{Calculate neighbor sums for a given matrix.}
\usage{
calculate_neighbour_sums(mat, cellneighbours)
}
\arguments{
\item{mat}{A matrix of counted molecules.}

\item{cellneighbours}{A dataframe containing cells and their respective neighbors.}
}
\value{
A matrix with neighbor sums.
}
\description{
Calculate neighbor sums for a given matrix.
}
