% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_Entropy.R
\name{combo_counts}
\alias{combo_counts}
\title{Convert cumulative counts in concentric polygons to combo polygon counts}
\usage{
combo_counts(mat)
}
\arguments{
\item{mat}{A matrix of cumulative counts.}
}
\value{
A matrix of combo polygon counts.
}
\description{
Convert cumulative counts in concentric polygons to combo polygon counts
}
