% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntropyMatrix_utils.R
\name{compute_cell_entropy}
\alias{compute_cell_entropy}
\title{Compute entropy for a given matrix}
\usage{
compute_cell_entropy(cell_segments_mat)
}
\arguments{
\item{cell_segments_mat}{A matrix to compute entropy for.}
}
\value{
A vector of entropy values.
}
\description{
Compute entropy for a given matrix
}
