% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_Entropy.R
\name{compute_entropy}
\alias{compute_entropy}
\title{Compute entropy for a given matrix}
\usage{
compute_entropy(mat)
}
\arguments{
\item{mat}{A matrix to compute entropy for.}
}
\value{
A vector of entropy values.
}
\description{
Compute entropy for a given matrix
}
