% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_MCC.R
\name{compute_mcc}
\alias{compute_mcc}
\title{Compute Matthews Correlation Coefficient}
\usage{
compute_mcc(table)
}
\arguments{
\item{table}{A 2x2 contingency table.}
}
\value{
The Matthews correlation coefficient calculated from the contingency table.
}
\description{
Compute Matthews Correlation Coefficient
}
