% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCCMatrix_utils.R
\name{count_molecules_from_SE}
\alias{count_molecules_from_SE}
\title{Count Molecules in SpatialExperiment}
\usage{
count_molecules_from_SE(me)
}
\arguments{
\item{me}{A SpatialExperiment object.}
}
\value{
An object with molecule counts for each cell.
}
\description{
Count Molecules in SpatialExperiment
}
