% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountsMatrix.R
\name{CountsMatrix}
\alias{CountsMatrix}
\title{Compute Counts Matrix for Specified Assay Name}
\usage{
CountsMatrix(me, assayName)
}
\arguments{
\item{me}{A data object that contains the input data.}

\item{assayName}{A character string specifying the assay name. It must be one
of "sub-sector", "sub-concentric", "sub-combo", "super-concentric", or
"super-combo".}
}
\value{
A matrix containing the counts data.
}
\description{
This function calculates the counts matrix for the specified assay name.
}
