% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountsMatrix.R
\name{CountsMatrix}
\alias{CountsMatrix}
\title{Extract Counts Matrix from Molecule Experiment based on Assay Type}
\usage{
CountsMatrix(me, assayName, ...)
}
\arguments{
\item{me}{A Molecule Experiment object.}

\item{assayName}{A character string indicating the assay type. Supported values include
"sub-sector", "sub-concentric", "sub-combo", "super-concentric", and "super-combo".}
}
\value{
A counts matrix corresponding to the specified assay type.
}
\description{
This function retrieves a counts matrix from a Molecule Experiment object based on the given assay type.
}
\examples{
\dontrun{
# Assuming `data_obj` is your Molecule Experiment object
cm = CountsMatrix(data_obj, assayName = "sub-sector")
}
}
