% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateFeatureData_utils.R
\name{create_sectors}
\alias{create_sectors}
\title{Create sectors from a given data frame}
\usage{
create_sectors(df)
}
\arguments{
\item{df}{A data frame to create sectors from.}
}
\value{
A data frame of sectors.
}
\description{
Create sectors from a given data frame
}
