% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_Entropy.R
\name{create_sectors_for_scaled}
\alias{create_sectors_for_scaled}
\title{Create sectors for scaled values}
\usage{
create_sectors_for_scaled(df, scale, k = 5)
}
\arguments{
\item{df}{A data frame to process.}

\item{scale}{A numeric scale factor.}

\item{k}{A numeric value, defaulting to 8.}
}
\value{
A data frame of sectors for scaled values.
}
\description{
Create sectors for scaled values
}
