% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_Entropy.R
\name{create_sectors_modified}
\alias{create_sectors_modified}
\title{Modify sectors based on a given data frame}
\usage{
create_sectors_modified(df)
}
\arguments{
\item{df}{A data frame to process.}
}
\value{
A modified data frame of sectors.
}
\description{
Modify sectors based on a given data frame
}
