% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountsEntropyPrefunctions.R
\name{delete_inner}
\alias{delete_inner}
\title{Delete Innermost Columns}
\usage{
delete_inner(m)
}
\arguments{
\item{m}{Matrix containing counts.}
}
\value{
A matrix without innermost polygon columns.
}
\description{
Deletes columns ending with '0', which represent the innermost polygon.
}
