% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntropyMatrix.R
\name{EntropyMatrix}
\alias{EntropyMatrix}
\title{Calculate the Entropy of a Counts Matrix}
\usage{
EntropyMatrix(me, assayNames, ...)
}
\arguments{
\item{me}{A Molecule Experiment object.}

\item{assayName}{A character string specifying the assay type. Supported values include
"sub-sector", "sub-concentric", "sub-combo", "super-concentric", and "super-combo".}
}
\value{
A matrix representing the entropy values corresponding to the given assay type.
}
\description{
This function computes the entropy of a counts matrix from a Molecule Experiment object based on the given assay type.
}
\examples{
\dontrun{
# Assuming `data_obj` is your Molecule Experiment object
em = EntropyMatrix(data_obj, assayName = "sub-sector")
}
}
