% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntropySummarizedExperiment.R
\name{EntropySummarizedExperiment}
\alias{EntropySummarizedExperiment}
\title{Create a SummarizedExperiment Object from a List of Data Frames}
\usage{
EntropySummarizedExperiment(df_list, me)
}
\arguments{
\item{df_list}{A list of data frames, each containing assay data.}

\item{me}{A Molecule Experiment object.}
}
\value{
A SummarizedExperiment object.
}
\description{
This function takes a list of data frames and a Molecule Experiment object and
constructs a SummarizedExperiment object with the specified data structure.
}
\examples{
\dontrun{
# Assuming `em_list_full` is your list of data frames and `me` is your Molecule Experiment object
se <- createSummarizedExperiment(df_list = em_list_full, me = me)
}
}
