% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_Entropy.R, R/utilities_MCC.R
\name{extract_boundaries}
\alias{extract_boundaries}
\title{Extract boundaries from a SpatialExperiment object.}
\usage{
extract_boundaries(me)

extract_boundaries(me)
}
\arguments{
\item{me}{A SpatialExperiment object.}
}
\value{
A data frame of extracted boundaries.

A dataframe containing boundaries.
}
\description{
Extract boundaries from a SpatialExperiment object.
}
