% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBoundaries_utils.R
\name{extract_boundaries}
\alias{extract_boundaries}
\title{Extract boundaries from a given object}
\usage{
extract_boundaries(me)
}
\arguments{
\item{me}{The object to extract boundaries from.}
}
\value{
A data frame of extracted boundaries.
}
\description{
Extract boundaries from a given object
}
