% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateFeatureData_utils.R,
%   R/MCCMatrix_utils.R
\name{extract_boundaries}
\alias{extract_boundaries}
\title{Extract boundaries from a given object}
\usage{
extract_boundaries(me)

extract_boundaries(me)
}
\arguments{
\item{me}{A SpatialExperiment object containing spatial and molecular data.}
}
\value{
A data frame of extracted boundaries.

A data frame with cell boundaries extracted from the SpatialExperiment object.
}
\description{
Extract boundaries from a given object

Extract Cell Boundaries from SpatialExperiment Object
}
