% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateFeatureData_utils.R
\name{extract_boundaries_and_centroids}
\alias{extract_boundaries_and_centroids}
\title{Extract boundaries and calculate centroids}
\usage{
extract_boundaries_and_centroids(me)
}
\arguments{
\item{me}{The object to process.}
}
\value{
A list containing a data frame of boundaries and a data frame of centroids.
}
\description{
Extract boundaries and calculate centroids
}
