% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_MCC.R
\name{extract_data_and_neighbors}
\alias{extract_data_and_neighbors}
\title{Extract boundaries and calculate neighbors for a SpatialExperiment object.}
\usage{
extract_data_and_neighbors(me)
}
\arguments{
\item{me}{A SpatialExperiment object.}
}
\value{
A list containing boundary data and neighbors.
}
\description{
Extract boundaries and calculate neighbors for a SpatialExperiment object.
}
