% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountsEntropyPrefunctions.R
\name{extract_sub_mat}
\alias{extract_sub_mat}
\title{Extract Sub Matrix}
\usage{
extract_sub_mat(mat, segment)
}
\arguments{
\item{mat}{Matrix from which to extract the sub-matrix.}

\item{segment}{Character string representing the segment of interest.}
}
\value{
A matrix corresponding to the specified segment.
}
\description{
Extracts a sub-matrix corresponding to a specific segment from a larger matrix.
}
