% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreMethodFunctions.R
\name{generate_scale_factors_all}
\alias{generate_scale_factors_all}
\title{Generate Scale Factors (All)}
\usage{
generate_scale_factors_all(k)
}
\arguments{
\item{k}{Numeric value for scaling.}
}
\value{
A vector of scale factors.
}
\description{
Generates scale factors for given value of k.
}
