% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_Entropy.R
\name{generate_scale_factors_all_outside}
\alias{generate_scale_factors_all_outside}
\title{Generate scale factors for all outside values of ks}
\usage{
generate_scale_factors_all_outside(ks)
}
\arguments{
\item{ks}{A numeric value.}
}
\value{
A vector of scale factors for outside values.
}
\description{
Generate scale factors for all outside values of ks
}
