% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreMethodFunctions.R
\name{generate_scale_factors_all_outside}
\alias{generate_scale_factors_all_outside}
\title{Generate Scale Factors (All Outside)}
\usage{
generate_scale_factors_all_outside(ks)
}
\arguments{
\item{ks}{Numeric value for scaling.}
}
\value{
A vector of scale factors for outer regions.
}
\description{
Generates scale factors for given value of ks.
}
