% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateFeatureData.R
\name{GenerateFeatureData}
\alias{GenerateFeatureData}
\title{Generate Feature Data Based on Assay Name}
\usage{
GenerateFeatureData(me, assayName)
}
\arguments{
\item{me}{A data object that contains the input data.}

\item{assayName}{A character string specifying the assay name. It must be one
of "sub-sector", "sub-concentric", "sub-combo", "super-concentric", or
"super-combo".}
}
\value{
A dataframe containing the processed feature data.
}
\description{
This function processes the input data according to the specified assay name
and returns the respective feature data.
}
