% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_MCC.R
\name{get_all_neighbors}
\alias{get_all_neighbors}
\title{Get All Neighbors in a Dataset}
\usage{
get_all_neighbors(df, threshold = 2)
}
\arguments{
\item{df}{A data frame with cell information, including 'segment_id' and 'sample_id'.}

\item{threshold}{Distance threshold to consider cells as neighbors (default is 2).}
}
\value{
A data frame listing pairs of cells and their neighbors.
}
\description{
Get All Neighbors in a Dataset
}
