% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_Entropy.R
\name{get_segments}
\alias{get_segments}
\title{Extract unique segments from matrix column names}
\usage{
get_segments(mat)
}
\arguments{
\item{mat}{A matrix to process.}
}
\value{
A character vector of unique segments.
}
\description{
Extract unique segments from matrix column names
}
