% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBoundaries.R
\name{loadBoundaries}
\alias{loadBoundaries}
\title{Load Boundaries to MoleculeExperiment Object}
\usage{
loadBoundaries(me, ...)
}
\arguments{
\item{me}{A MoleculeExperiment (ME) object}

\item{...}{arguments passed to \code{\link{GenerateFeatureData}}}
}
\value{
An enriched MoleculeExperiment (ME) object that includes boundary
information for the supported assays.
}
\description{
This function takes a \linkS4class{MoleculeExperiment} (ME) object as input
and enriches it with boundary data from a variety of assays. The currently
supported assays are:
\itemize{
\item Sub-sector polygons
\item Sub-concentric polygons
\item Super-sector polygons
\item Super-concentric polygons
}

The function internally processes each assay, extracts respective boundary
information, and updates the original ME object with this new data.
}
\examples{
data(example_me)
me <- loadBoundaries(me)
me
}
