% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntropySummarizedExperiment_utils.R
\name{make_assay_data}
\alias{make_assay_data}
\title{make assay data}
\usage{
make_assay_data(df_list, concatenateFeatures = FALSE)
}
\arguments{
\item{df_list}{A list of data frames}

\item{concatenateFeatures}{logical whether to concatenate features (default FALSE)}
}
\value{
if concatenateFeatures == TRUE, A single data frame, otherwise a list
containing data frames
}
\description{
make assay data
}
