% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCCMatrix.R
\name{MCCMatrix}
\alias{MCCMatrix}
\title{Comprehensive Function to Compute MCC Matrix from SpatialExperiment}
\usage{
MCCMatrix(me)
}
\arguments{
\item{me}{A SpatialExperiment object.}
}
\value{
A sparse matrix of Matthews correlation coefficients.
}
\description{
This function wraps several processing steps including extracting cell boundaries,
counting molecules, calculating neighbor sums, ordering and sparsifying the matrix,
and computing the MCC matrix.
}
\examples{
# Assuming 'small_me' is a SpatialExperiment object
# not run
# data(small_me, package = "MoleculeExperiment")
# mcc_matrix <- MCCMatrix(small_me)
}
