% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_MCC.R
\name{order_and_sparse_matrix}
\alias{order_and_sparse_matrix}
\title{Order and Convert Matrix to Sparse Format}
\usage{
order_and_sparse_matrix(mat)
}
\arguments{
\item{mat}{A matrix to be ordered and converted.}
}
\value{
A sparse matrix in ordered format.
}
\description{
Order and Convert Matrix to Sparse Format
}
